backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pytesseract>=0.3.10
opencv-python>=4.8.1.78
pdf2image>=1.16.3
PyPDF2>=3.0.1
Pillow>=10.2.0
python-docx>=1.1.0
numpy>=1.26.0
